07 nov merge andre (#59)
* Remove events first_descendant and last_ancestors

* Tests, poset rename, and plays

* Internal transactions for peer prototype

* Signature verification fail patch

* EvenBlock signature debugging

* SetWireInfo before signing

* SetWireInfoAndSign reorder

* Log sent events for signature comparison

* InternalTransactions from wireInfo

* Compare wireInfo and &out.response

* WireEvent Body comparison

* Sig compare

* Don't sign and insert

* Test send vs receive on wireEvent

* RPC target

* Exclude pubkey from peers list

* conflict resolution in merging in  Andre's master

+++ src/poset/debug.go
@@ -10,8 +10,8 @@ import (
 func (p *Poset) PrintStat() {
 	fmt.Println("****Known events:");
 	for pid_id, index := range p.Store.KnownEvents() {
-	    fmt.Println("    index=", index, " peer=", p.Participants.ById[pid_id].NetAddr,
-			" pubKeyHex=", p.Participants.ById[pid_id].PubKeyHex)
+		fmt.Println("    index=", index, " peer=", p.Participants.ById[int64(pid_id)].NetAddr,
+			" pubKeyHex=", p.Participants.ById[int64(pid_id)].PubKeyHex)
 	}
 }
